DDR4: Honour ACT->REF distance.
This is a safeguard when e.g. the following sequence of commands occurs
with minimal distance:
ACT(x), RDA(x), REF

In this case RDA will be issued as early as possible for minimum data arrival
delay, but not finish until after tRC. The existing distance between RDA and
REF is insufficient to guarantee compliant timing, as generally nRAS > nRTP+nRP.
This patch adds a minimal distance between ACT and REF, mirroring the ACT-ACT
distance at the bank level.

Signed-off-by: Roy Spliet <rs855@cam.ac.uk>