Example of a debug flag
Example on how to print things only when a certain debug flag is specified.
The current implementation is not very efficient,
 as there is a call each time to check the config class whether a string contains a substring.
Debug can be turned off by not defining DEBUG.
A good suggestion would be to make a new make target 'debug'
so the compiler turns DEBUG on or off based on compilation flags.
Another suggestion is to make the DEBUG flag printing more efficient.